Address authentication issues with Cognito
When performing requests that require authentication, you have to get the session from the cognito user. This will initialize and setup everyting needed to perform the request. If this is not done, you recieve a `User not authenticated` error.

This also cleans up the success case in the fetch attributes where it was setting the wrong state

ps-id: EC6FE484-FA57-4A36-896F-65DAA6C4C5F0